src/agent.py
"""
Enterprise Agent - Main Implementation

This module contains the core agent logic for the capstone project.
"""

import os
import logging
from typing import Dict, List, Any, Optional
from dataclasses import dataclass
from dotenv import load_dotenv # type: ignore

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)


@dataclass
class AgentConfig:
    """Configuration for the enterprise agent."""
    api_key: str
    model: str = "gpt-4"
    temperature: float = 0.7
    max_tokens: int = 2000
    
    @classmethod
    def from_env(cls) -> "AgentConfig":
        """Create configuration from environment variables."""
        api_key = os.getenv("API_KEY", "")
        model = os.getenv("MODEL", "gpt-4")
        temperature = float(os.getenv("TEMPERATURE", "0.7"))
        max_tokens = int(os.getenv("MAX_TOKENS", "2000"))
        
        return cls(
            api_key=api_key,
            model=model,
            temperature=temperature,
            max_tokens=max_tokens
        )


class EnterpriseAgent:
    """
    Main Enterprise Agent class.
    
    This agent is designed to solve [TODO: describe your specific use case].
    """
    
    def __init__(self, config: AgentConfig):
        """
        Initialize the enterprise agent.
        
        Args:
            config: Agent configuration
        """
        self.config = config
        self.tools = []
        self.conversation_history = []
        
        logger.info("Enterprise Agent initialized")
        
    def add_tool(self, tool: Any) -> None:
        """
        Add a tool to the agent's toolkit.
        
        Args:
            tool: Tool instance to add
        """
        self.tools.append(tool)
        logger.info(f"Added tool: {tool.__class__.__name__}")
        
    def process(self, input_data: str) -> Dict[str, Any]:
        """
        Process input and generate agent response.
        
        Args:
            input_data: Input text or data to process
            
        Returns:
            Dictionary containing the agent's response and metadata
        """
        logger.info(f"Processing input: {input_data[:100]}...")
        
        # TODO: Implement your agent logic here
        # This is where you'll integrate with LLMs, use tools, and orchestrate the workflow
        
        result = {
            "status": "success",
            "response": "Agent processing not yet implemented",
            "input": input_data,
            "tools_used": [],
            "metadata": {}
        }
        
        return result
    
    def run(self) -> None:
        """
        Run the agent in interactive mode.
        """
        logger.info("Starting agent in interactive mode...")
        print("Enterprise Agent Ready!")
        print("Type 'exit' to quit\n")
        
        while True:
            try:
                user_input = input("You: ").strip()
                
                if user_input.lower() in ['exit', 'quit', 'q']:
                    print("Goodbye!")
                    break
                    
                if not user_input:
                    continue
                
                result = self.process(user_input)
                print(f"Agent: {result['response']}\n")
                
            except KeyboardInterrupt:
                print("\nGoodbye!")
                break
            except Exception as e:
                logger.error(f"Error processing input: {e}")
                print(f"Error: {e}\n")


def main():
    """Main entry point for the agent."""
    try:
        # Load configuration
        config = AgentConfig.from_env()
        
        # Create and run agent
        agent = EnterpriseAgent(config)
        agent.run()
        
    except Exception as e:
        logger.error(f"Fatal error: {e}")
        raise


if __name__ == "__main__":
    main()
