requirements.txt
# Core dependencies
python-dotenv==1.0.0

# LLM and Agent frameworks
openai==1.12.0
anthropic==0.18.1
langchain==0.1.9
langchain-community==0.0.24
langchain-openai==0.0.6

# Alternative: LlamaIndex
# llama-index==0.10.12

# HTTP and API
requests==2.31.0
httpx==0.26.0

# Data processing
pandas==2.2.0
numpy==1.26.3

# Logging and monitoring
loguru==0.7.2

# Testing
pytest==8.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0

# Code quality
black==24.1.1
flake8==7.0.0
mypy==1.8.0

# Utilities
pydantic==2.6.1
pyyaml==6.0.1
